requirements.txt
torch
tensorflow
onnx
tf2onnx
onnxruntime-gpu